refactor: rename `mainbanner.tsx` to `BannerHero.tsx` to keep consistency
